% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_annotator.R
\name{active_annotator}
\alias{active_annotator}
\title{Text Annotation Using Active Learning}
\usage{
active_annotator(
  data,
  text_col = "text",
  y_col = "y",
  doc_id_col = "doc_id",
  method = "lasso",
  n_batches = 5,
  batch_size = 10,
  rand_minority_min_n = 100,
  n_cores = NULL
)
}
\arguments{
\item{data}{A dataframe with a 'text' column and 'doc_id' column.}

\item{text_col}{A string with the name of the column containing document 
text. Defaults to 'text'}

\item{y_col}{A string with the name of the column containing document 
text. Defaults to 'y'}

\item{doc_id_col}{A string with the name of the column containing document
ids. Defaults to 'doc_id'}

\item{method}{'lasso' samples unlabeled documents with a predicted 
response nearest to .5 from a lasso regression classifier. 'svm' samples 
unlabeled documents closest to the class-separating hyperplane of a 
support vector machine. 'rf' samples unlabeled documents based on the
vote entropy of all trees in a random forest classifier.}

\item{n_batches}{An integer for the number of total batches to label.
Defaults to 10..}

\item{batch_size}{An integer for the number of documents to sample at 
each iteration of the active learning algorithm. Defaults to 10.}

\item{rand_minority_min_n}{An integer for the minimum number of documents
in the minority class needed to begin active learning. If the minority
class has fewer observations than `rand_minority_min_n`, documents will
be sampled randomly. Defaults to 100.}

\item{n_cores}{An integer for the number of cores to use in the estimation 
of the model.}
}
\value{
A new data frame object derived from the data frame passed into
the the 'data' argument. This new data frame includes new annotations from
the annotation session in the column specified in the 'y_col' argument. 
A new column 'batch' has been added to specify the batch at which each 
new document was labeled, and a new column 'active' has been added to 
specify  whether a newly-labeled document was sampled actively or 
passively (randomly).
}
\description{
This function provides a simple R console interface for annotation of documents using active learning. It can sample texts for annotation using a variety of active learning methods. This function returns a data frame with annotations and their resepective 'batch_id' and a boolean flag 'active' to identify which batches were sampled actively. This flag helps identify randomly-sampled data which can provide an unbiased estimate of generalization error.
}
\examples{
data <- read.csv('../example_data/hate_speech.csv', stringsAsFactors=FALSE)
data_annotated <- active_annotator(data, text_col='comment_text', y_col='obscene', doc_id_col='id', n_batches=10, batch_size=5, method='lasso')

}
\author{
Blake Miller
}
\keyword{active}
\keyword{annotation}
\keyword{learning}
\keyword{supervised}
\keyword{text}
