% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textannotation_random.R
\name{textannotation_random}
\alias{textannotation_random}
\title{Text Annotation Using Random Sampling}
\usage{
textannotation_random(
  data,
  text_col = "text",
  y_col = "y",
  doc_id_col = "doc_id",
  n_docs = 100
)
}
\arguments{
\item{data}{A dataframe with a 'text' column and 'doc_id' column.}

\item{text_col}{A string with the name of the column containing document text. Defaults to 'text'}

\item{y_col}{A string with the name of the column containing document text. Defaults to 'y'}

\item{doc_id_col}{A string with the name of the column containing document ids. Defaults to 'doc_id'}

\item{n_docs}{An integer for the number of documents to label in the session.}
}
\value{
An data frame object derived from the data frame specified in
the 'data' parameter where the new annotations have been added to the
the column specified in 'y_col', a new column 'active' has been added
to specifying whether or not a newly-labeled document was sampled 
actively or passively (randomly).
}
\description{
This function supports annotation of documents using random sampling.
}
\examples{
data <- read.csv('hate_speech.csv', stringsAsFactors=FALSE)
data_annotated <- textannotation_random(data, text_col='comment_text', y_col='toxic', doc_id_col='id', n_docs=100)
}
\author{
Blake Miller
}
\keyword{active}
\keyword{annotation}
\keyword{learning}
\keyword{supervised}
\keyword{text}
