% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textannotation_rf.R
\name{textannotation_rf}
\alias{textannotation_rf}
\title{Text Annotation Using Query by Committee via Random Forests}
\usage{
textannotation_rf(
  data,
  text_col = "text",
  y_col = "y",
  doc_id_col = "doc_id",
  n_batches = 5,
  batch_size = 20,
  rand_minority_min_n = 100
)
}
\arguments{
\item{data}{A dataframe with a 'text' column and 'doc_id' column.}

\item{text_col}{A string with the name of the column containing document 
text. Defaults to 'text'}

\item{y_col}{A string with the name of the column containing document 
text. Defaults to 'y'}

\item{doc_id_col}{A string with the name of the column containing document
ids. Defaults to 'doc_id'}

\item{n_batches}{An integer for the number of total batches to label.
Defaults to 10..}

\item{batch_size}{An integer for the number of documents to sample at 
each iteration of the active learning algorithm. Defaults to 10.}

\item{rand_minority_min_n}{An integer for the minimum number of documents
in the minority class needed to begin active learning. If the minority
class has fewer observations than `rand_minority_min_n`, documents will
be sampled randomly. Defaults to 100.}
}
\value{
A new data frame object derived from the data frame passed into
the the 'data' argument. This new data frame includes new annotations from
the annotation session in the column specified in the 'y_col' argument. 
A new column 'batch' has been added to specify the batch at which each 
new document was labeled, and a new column 'active' has been added to 
specify  whether a newly-labeled document was sampled actively or 
passively (randomly).
}
\description{
This function supports annotation of documents using a "query by committee" approach using vote entropy from tree classifiers in a random forest ensemble. This algorithm samples a batch of unlabeled documents with a forest vote entropy closest to 1 for annotation via the R console. This function returns a data frame with annotations and their resepective batch_id and a boolean flag for whether or not the documents were sampled actively (or randomly).
}
\examples{
data <- read.csv('../example_data/hate_speech.csv', stringsAsFactors=FALSE)
data_annotated <- textannotation_rf(data, text_col='comment_text', y_col='obscene', doc_id_col='id', n_batches=5, batch_size=20)
}
\author{
Blake Miller
}
\keyword{active}
\keyword{annotation}
\keyword{learning}
\keyword{supervised}
\keyword{text}
