% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textannotation_lasso.R
\name{textannotation_lasso}
\alias{textannotation_lasso}
\title{Text Annotation Using Uncertainty Sampling via Lasso Regression}
\usage{
textannotation_lasso(
  data,
  text_col = "text",
  y_col = "y",
  doc_id_col = "doc_id",
  n_batches = 10,
  batch_size = 10,
  rand_minority_min_n = 100,
  n_cores = 2
)
}
\arguments{
\item{data}{A dataframe with a 'text' column and 'doc_id' column.}

\item{text_col}{A string with the name of the column containing document 
text. Defaults to 'text'}

\item{y_col}{A string with the name of the column containing document 
text. Defaults to 'y'}

\item{doc_id_col}{A string with the name of the column containing document
ids. Defaults to 'doc_id'}

\item{n_batches}{An integer for the number of total batches to label.
Defaults to 10..}

\item{batch_size}{An integer for the number of documents to sample at 
each iteration of the active learning algorithm. Defaults to 10.}

\item{rand_minority_min_n}{An integer for the minimum number of documents
in the minority class needed to begin active learning. If the minority
class has fewer observations than `rand_minority_min_n`, documents will
be sampled randomly. Defaults to 100.}

\item{n_cores}{An integer for the number of cores to use in the estimation 
of the lasso model.}
}
\value{
A new data frame object derived from the data frame passed into
the the 'data' argument. This new data frame includes new annotations from
the annotation session in the column specified in the 'y_col' argument. 
A new column 'batch' has been added to specify the batch at which each 
new document was labeled, and a new column 'active' has been added to 
specify  whether a newly-labeled document was sampled actively or 
passively (randomly).
}
\description{
This function supports annotation of documents using an uncertainty sampling approach using predicted response of lasso regression as an uncertainty measure. This algorithm samples a batch of unlabeled documents with a predicted probability nearest to .5 for annotation via the R console. This function returns a data frame with annotations and their resepective batch_id and a boolean flag for whether or not the documents were sampled actively (or randomly).
}
\examples{
data <- read.csv('wikipedia_hate_speech.csv', stringsAsFactors=FALSE)
data_annotated <- textannotation_lasso(data, text_col='comment_text', y_col='toxic', doc_id_col='id', n_batches=100, batch_size=5)
}
\author{
Blake Miller
}
\keyword{active}
\keyword{annotation}
\keyword{learning}
\keyword{supervised}
\keyword{text}
